Fix "Error opening terminal: xterm-256color." error when launching
The snap lacks files from the ncurses-base package, which was not
shipped in the snap nor the ncurses-launch stage snap, causing the
error.

This patch also migrates to the new implementation of ncurses-launch
launcher, which dumps the source repository instead of using the built
stage-snap, which is no longer unsupported.

Fixes #23.

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>